// src/pages/HealthProfessional/Onboarding/SecondStep/utils/types/healthProfessionalPreferences.ts
export type HealthProfessionalPreferencesInputs = {
  clinicalTrialsInfo: boolean
  geneticsKnowledge: boolean
  collaborateWithSpecialists: boolean
  createHealthNetwork: boolean
  teamworkSkills: boolean
  analyzeClinicalCases: boolean
  latestResearchArticles: boolean
  webinarsParticipation: boolean
  applyAsSpeaker: boolean
  patientHistoryAndSymptoms: boolean
  treatmentGuide: boolean
  improveKnowledgeCertification: boolean
  allOptions: boolean
  otherOption: string
  healthProfessionalId: number
}